pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Aeropulse"
version = "0.1.1"
readme = "README.md"
authors = [{ name = "veziriii"}]
requires-python = ">=3.10"

dependencies = [
  "requests",
  "python-dotenv",
  "pymongo>=4.6",
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "h3>=3.7",
  "pandas>=2.2",
  "pyarrow>=15",
  "dask[dataframe]>=2024.5.0",
  "matplotlib",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-cov",
  "black",
  "isort",
  "mypy",
  "ruff",
  "types-requests",
]


[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]   # find all subpackages under src
